requirements.txt
numpy
tensorflow-gpu==1.10.1
